requirements.txt
streamlit
Pillow
pyzipper